requirements.txt
Flask
openai
sqlparse
Flask-CORS
python-dotenv
